+++ express-zod-api/tests/endpoints-factory.spec.ts
@@ -99,11 +99,9 @@ describe("EndpointsFactory", () => {
       expect(factory["middlewares"]).toStrictEqual([]);
       expect(factory["resultHandler"]).toStrictEqual(resultHandlerMock);
       expect(newFactory["middlewares"].length).toBe(1);
-      expect(newFactory["middlewares"][0].getSchema()).toBeInstanceOf(
-        z.ZodObject,
-      );
+      expect(newFactory["middlewares"][0].schema).toBeInstanceOf(z.ZodObject);
       expect(
-        (newFactory["middlewares"][0].getSchema() as z.AnyZodObject).shape,
+        (newFactory["middlewares"][0].schema as z.AnyZodObject).shape,
       ).toEqual({});
       const { output: options } = await testMiddleware({
         middleware: newFactory["middlewares"][0],
@@ -129,11 +127,9 @@ describe("EndpointsFactory", () => {
           provider: (req) => ({ result: req.body.test }),
         });
         expect(newFactory["middlewares"].length).toBe(1);
-        expect(newFactory["middlewares"][0].getSchema()).toBeInstanceOf(
-          z.ZodObject,
-        );
+        expect(newFactory["middlewares"][0].schema).toBeInstanceOf(z.ZodObject);
         expect(
-          (newFactory["middlewares"][0].getSchema() as z.AnyZodObject).shape,
+          (newFactory["middlewares"][0].schema as z.AnyZodObject).shape,
         ).toEqual({});
         const {
           output: options,